+++ common/src/main/java/traben/entity_texture_features/mixin/entity/renderer/feature/MixinShoulderParrotFeatureRenderer.java
@@ -6,10 +6,8 @@ import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.RenderLayerParent;
 import net.minecraft.client.renderer.entity.layers.ParrotOnShoulderLayer;
 import net.minecraft.client.renderer.entity.layers.RenderLayer;
-import net.minecraft.client.renderer.entity.state.PlayerRenderState;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EntitySpawnReason;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.animal.Parrot;
 import net.minecraft.world.entity.player.Player;
@@ -23,10 +21,14 @@ import traben.entity_texture_features.utils.ETFEntity;
 
 import java.util.Optional;
 
-@Mixin(ParrotOnShoulderLayer.class)
 #if MC > MC_21
+import net.minecraft.client.renderer.entity.state.PlayerRenderState;
+import net.minecraft.world.entity.EntitySpawnReason;
+
+@Mixin(ParrotOnShoulderLayer.class)
 public abstract class MixinShoulderParrotFeatureRenderer extends RenderLayer<PlayerRenderState, PlayerModel> {
 #else
+@Mixin(ParrotOnShoulderLayer.class)
 public abstract class MixinShoulderParrotFeatureRenderer<T extends Player> extends RenderLayer<T, PlayerModel<T>> {
 #endif