+++ fabric/src/main/java/traben/entity_texture_features/fabric/mixin/MixinArmorFeatureRenderer.java
@@ -5,16 +5,20 @@ package traben.entity_texture_features.fabric.mixin;
 
 #if MC < MC_21_2
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.RenderLayerParent;
+import net.minecraft.client.renderer.entity.layers.HumanoidArmorLayer;
+import net.minecraft.client.renderer.entity.layers.RenderLayer;
 import net.minecraft.resources.ResourceLocation;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import net.minecraft.util.FastColor;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ArmorMaterial;
 import net.minecraft.world.item.armortrim.ArmorTrim;
 import net.minecraft.world.item.ArmorItem;
-import net.minecraft.world.item.equipment.ArmorMaterial;
-import net.minecraft.world.item.equipment.trim.ArmorTrim;
+
 import org.spongepowered.asm.mixin.injection.ModifyArg;
 import traben.entity_texture_features.features.texture_handlers.ETFArmorHandler;
 import org.spongepowered.asm.mixin.Unique;